# core: batch get region size