--- a/pkg/core/region_test.go
+++ b/pkg/core/region_test.go
@@ -706,7 +706,7 @@ func TestGetRegionSizeByRange(t *testing.T) {
 	totalSize := regions.GetRegionSizeByRange([]byte(""), []byte(""))
 	require.Equal(t, int64(nums*10), totalSize)
 	for i := 1; i < 10; i++ {
-		verifyNum := int(nums / i)
+		verifyNum := nums / i
 		endKey := fmt.Sprintf("%20d", verifyNum)
 		totalSize := regions.GetRegionSizeByRange([]byte(""), []byte(endKey))
 		require.Equal(t, int64(verifyNum*10), totalSize)