@nolouch 2023-11-07
I added a benchmark and ran it in my local MBP, it shows about 1000% improvement.

master
```

go test -benchmem -run=^$ -bench ^BenchmarkRandomSetRegionWithGetRegionSizeByRange$ github.com/tikv/pd/pkg/core --tags=intest

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/core
BenchmarkRandomSetRegionWithGetRegionSizeByRange-12    	   88456	     12543 ns/op	       0 B/op	       0 allocs/op


Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkRandomSetRegionWithGetRegionSizeByRangeParallel$ github.com/tikv/pd/pkg/core --tags=intest

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/core
BenchmarkRandomSetRegionWithGetRegionSizeByRangeParallel-12    	   22857	     58492 ns/op	     755 B/op	      15 allocs/op
PASS
ok  	github.com/tikv/pd/pkg/core	14.837s

```
this branch:

```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkRandomSetRegionWithGetRegionSizeByRange$ github.com/tikv/pd/pkg/core --tags=intest

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/core
BenchmarkRandomSetRegionWithGetRegionSizeByRange-12    	  601136	      1928 ns/op	    4878 B/op	      35 allocs/op
PASS



Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkRandomSetRegionWithGetRegionSizeByRangeParallel$ github.com/tikv/pd/pkg/core --tags=intest

goos: darwin
goarch: arm64
pkg: github.com/tikv/pd/pkg/core
BenchmarkRandomSetRegionWithGetRegionSizeByRangeParallel-12    	  241456	      5818 ns/op	   12736 B/op	     103 allocs/op
PASS
ok  	github.com/tikv/pd/pkg/core	16.234s



```